Replace overworld action routing with typed tools
Collapse tool selection and secondary argument requests into one typed overworld decision. Preserve deterministic navigation and emulator services while filtering available actions from the current game state.
