--- agent/subflows/overworld_handler/nodes/navigate/schemas.py
@@ -1,12 +1,0 @@
-"""Data models for navigate in the overworld subflow."""
-
-from pydantic import BaseModel
-
-from common.schemas import Coords
-
-
-class NavigationResponse(BaseModel):
-    """The response from the overworld navigation prompt."""
-
-    thoughts: str
-    coords: Coords
