--- agent/subflows/overworld_handler/nodes/press_buttons/schemas.py
@@ -1,12 +1,0 @@
-"""Data models for press buttons in the overworld subflow."""
-
-from pydantic import BaseModel
-
-from common.enums import Button
-
-
-class PressButtonsResponse(BaseModel):
-    """The response from the overworld button presser prompt."""
-
-    thoughts: str
-    buttons: Button | list[Button]
