--- agent/subflows/overworld_handler/nodes/sokoban_solver/node.py
@@ -1,40 +1,0 @@
-"""Sokoban solver node for the overworld subflow."""
-
-from typing import TYPE_CHECKING
-
-from junjo import Node
-from loguru import logger
-
-from agent.subflows.overworld_handler.nodes.sokoban_solver.service import SokobanSolverService
-from agent.subflows.overworld_handler.state import OverworldHandlerStore
-
-if TYPE_CHECKING:
-    from emulator.emulator import YellowLegacyEmulator
-
-
-class SokobanSolverNode(Node[OverworldHandlerStore]):
-    """Solve the Sokoban puzzle."""
-
-    def __init__(self, emulator: YellowLegacyEmulator) -> None:
-        """Initialize the sokoban solver node."""
-        self.emulator = emulator
-        super().__init__()
-
-    async def service(self, store: OverworldHandlerStore) -> None:
-        """The service for the node."""
-        logger.info("Using the Sokoban solver...")
-
-        state = await store.get_state()
-        if state.rolling_memory is None:
-            raise ValueError("Rolling memory is not set")
-        if state.current_map is None:
-            raise ValueError("Current map is not set")
-
-        service = SokobanSolverService(
-            emulator=self.emulator,
-            current_map=state.current_map,
-            rolling_memory=state.rolling_memory,
-        )
-        await service.solve()
-
-        await store.set_rolling_memory(state.rolling_memory)
