--- agent/subflows/overworld_handler/nodes/swap_first_pokemon/schemas.py
@@ -1,13 +1,0 @@
-"""Data models for swap first Pokémon in the overworld subflow."""
-
-from pydantic import BaseModel, Field
-
-
-class SwapPokemonError(Exception):
-    """An error that occurs when swapping a Pokemon."""
-
-
-class SwapFirstPokemonResponse(BaseModel):
-    """The response from the overworld swap first Pokemon prompt."""
-
-    index: int = Field(ge=0, le=5)
