--- agent/subflows/overworld_handler/nodes/use_item/service.py
@@ -1,107 +1,0 @@
-"""Business logic for use item in the overworld subflow."""
-
-from typing import TYPE_CHECKING
-
-from loguru import logger
-
-from agent.subflows.overworld_handler.nodes.use_item.prompts import USE_ITEM_PROMPT
-from agent.subflows.overworld_handler.nodes.use_item.schemas import UseItemError, UseItemResponse
-from common.enums import Button
-from llm.service import OpenAILLMService
-
-if TYPE_CHECKING:
-    from emulator.emulator import YellowLegacyEmulator
-    from memory.rolling_memory import RollingMemory
-
-
-class UseItemService:
-    """A service that uses an item from the inventory."""
-
-    llm_service = OpenAILLMService()
-
-    def __init__(
-        self,
-        rolling_memory: RollingMemory,
-        emulator: YellowLegacyEmulator,
-    ) -> None:
-        """Initialize the use item service."""
-        self.rolling_memory = rolling_memory
-        self.emulator = emulator
-
-    async def use_item(self) -> RollingMemory:
-        """Use an item from the inventory."""
-        try:
-            index = await self._get_item_index()
-            await self._use_item(index)
-        except Exception as e:  # noqa: BLE001
-            logger.warning(f"Error in the use item response. Skipping. {e}")
-            self.rolling_memory.add_memory(
-                content=f"I failed to use an item from my inventory. {e}",
-            )
-        return self.rolling_memory
-
-    async def _get_item_index(self) -> int:
-        """Get the index of the item to use."""
-        game_state = await self.emulator.get_game_state()
-        last_memory = str(self.rolling_memory.current_block)
-        inventory_str = "\n".join(
-            f"[{i}] {item.name}" for i, item in enumerate(game_state.inventory.items)
-        )
-        prompt = USE_ITEM_PROMPT.format(thought=last_memory, inventory=inventory_str)
-        response = await self.llm_service.get_llm_response_pydantic(
-            messages=prompt,
-            schema=UseItemResponse,
-        )
-        return response.index
-
-    async def _use_item(self, item_index: int) -> None:
-        """Use the item at the given index."""
-        game_state = await self.emulator.get_game_state()
-        if game_state.is_text_on_screen():
-            raise UseItemError("Can't use an item in a non-overworld state.")
-
-        # Splitting into sub-steps for easier debugging. Otherwise the various game states become
-        # too difficult to keep track of.
-        await self._open_start_menu()
-        await self._open_item_menu()
-        await self._select_item(item_index)
-
-    async def _open_start_menu(self) -> None:
-        """Open the start menu."""
-        await self.emulator.press_button(Button.START)
-        game_state = await self.emulator.get_game_state()
-        screen_text = game_state.screen.text
-        if "POKéDEX" not in screen_text and "POKéMON" not in screen_text:
-            raise UseItemError("Failed to open the START menu.")
-
-    async def _open_item_menu(self) -> None:
-        """Open the ITEM menu."""
-        game_state = await self.emulator.get_game_state()
-        item_menu_position = 2
-        idx_diff = game_state.screen.menu_item_index - item_menu_position
-        await self._move_cursor(idx_diff)
-
-        screen_text = (await self.emulator.get_game_state()).screen.text
-        if "▶ITEM" not in screen_text:
-            raise UseItemError("Failed to open the ITEM menu.")
-        await self.emulator.press_button(Button.A)
-
-    async def _select_item(self, item_index: int) -> None:
-        """Move the cursor to the item at the given index."""
-        screen = (await self.emulator.get_game_state()).screen
-        idx_diff = screen.menu_item_index + screen.list_scroll_offset - item_index
-        await self._move_cursor(idx_diff)
-        await self.emulator.press_button(Button.A)  # Select the item.
-        await self.emulator.press_button(Button.A)  # Use the item.
-
-    async def _move_cursor(self, step: int) -> None:
-        """Move a vertical menu cursor by a signed number of steps.
-
-        Positive values move up; negative values move down.
-        """
-        if step > 0:
-            for _ in range(step):
-                await self.emulator.press_button(Button.UP)
-        elif step < 0:
-            for _ in range(-step):
-                await self.emulator.press_button(Button.DOWN)
